+++ inst/include/jsonify/to_json/writers/simple.hpp
@@ -150,8 +150,6 @@ namespace simple {
       
     } else if ( factors_as_string && Rf_isFactor( iv ) ) {
       
-      //Rcpp::Rcout << "factors 1" << std::endl;
-      
       Rcpp::CharacterVector lvls = iv.attr( "levels" );
       if (lvls.length() == 0 && iv.length() == 0 ) {
         writer.StartArray();
@@ -211,7 +209,7 @@ namespace simple {
       write_value( writer, sv, row );
       
     } else if ( factors_as_string && Rf_isFactor( iv ) ) {
-      //Rcpp::Rcout << "factors 2" << std::endl;
+      
       Rcpp::StringVector lvls = iv.attr( "levels" );
       
       if (lvls.length() == 0 && iv.length() == 0 ) {